♻️ [Refactoring]: #455 Phase 10 byte_preservation カテゴリを tests/byte_preservation.rs に分離
NUL / 高位 / 非 UTF8 バイト保存のテスト 3 件を
literal_string/tests/byte_preservation.rs に切り出し、tests.rs ハブに
`mod byte_preservation;` 宣言を追加。
