+++ rust/crates/core/src/lexer/literal_string/tests.rs
@@ -1,6 +1,7 @@
 use super::*;
 
 mod basic;
+mod byte_preservation;
 mod eol_normalize;
 mod error;
 mod escape;
@@ -9,34 +10,6 @@ mod line_continuation;
 mod nest;
 mod octal;
 mod unknown_escape;
-
-// ========================================================================
-// Phase 11-J: 非 ASCII / NUL / 高位バイト保持
-// ========================================================================
-
-#[test]
-fn read_literal_string_preserves_nul_byte() {
-    // b"(\x00)" で NUL バイトをそのまま保持し pos == 3 を返すことを確認する
-    let mut lexer = Lexer::new(b"(\x00)");
-    assert_eq!(lexer.read_literal_string(), Some(b"\x00".to_vec()));
-    assert_eq!(lexer.position(), 3);
-}
-
-#[test]
-fn read_literal_string_preserves_high_byte() {
-    // b"(\xFF)" で 0xFF バイトをそのまま保持し pos == 3 を返すことを確認する
-    let mut lexer = Lexer::new(b"(\xFF)");
-    assert_eq!(lexer.read_literal_string(), Some(b"\xFF".to_vec()));
-    assert_eq!(lexer.position(), 3);
-}
-
-#[test]
-fn read_literal_string_preserves_non_utf8_sequence() {
-    // b"(\x80\xC0)" で非 UTF-8 連続バイト列をそのまま保持し pos == 4 を返すことを確認する
-    let mut lexer = Lexer::new(b"(\x80\xC0)");
-    assert_eq!(lexer.read_literal_string(), Some(b"\x80\xC0".to_vec()));
-    assert_eq!(lexer.position(), 4);
-}
 
 // ========================================================================
 // decode_escape 純関数の単体テスト
